add switch to activate "Throw Balls in Player" mode permanently and define the default ball size for this mode add an "Update All" button to the image manager
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@2480 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ codeview.h
@@ -276,7 +276,7 @@ public:
 	bool DisplayAutoComplete;
 	void GetMembers(vector<UserData>* ListIn, const string &StrIn);
 	//TODO: int TabStop;
-	int DisplayAutoCompleteLength;
+	size_t DisplayAutoCompleteLength;
 	bool DwellDisplay;
 	bool DwellHelp;
 	int DwellDisplayTime;
